cleanup: rename folder
