+++ packages/client/src/components/documents-table/cells/amount.tsx
